feat: relocate test file
